Improved gate macro
Can now define gates that take multiple qubits; are or are not
self-adjoint. It's a little awkward we must still manually call
`m.add_class::<_>` in the module definition for each gate type, but I
don't yet see a way around this. Expanding the macro within the module
would leave the gates undefined outside of it.